Function Library

Function Library

Schrodinger.ptrace — Method.

ptrace(ρ, out)

Compute the partial trace of a Density matrix or Operator ρ by tracing out the subsystems specified by out. Multiple subsystems can be traced out by passing a sorted tuple of subsystem indices.

Example

Φ₊ = normalize!(basis(2,0)⊗basis(2,0) + basis(2,1)⊗basis(2,1)) # Bell pair
Ψ₊ = normalize!(basis(2,0)⊗basis(2,1) + basis(2,1)⊗basis(2,0)) # Bell pair
ρ  = 0.25 * Density(Φ₊) + 0.75 * Density(Ψ₊) # density matrix
ptrace(ρ,2) # trace out qubit 2
# output
2×2 Schrodinger.Density{Array{Float64,2},1} with space dimensions 2:
 0.5  0.0
 0.0  0.5

source

Schrodinger.ptrace — Method.

ptrace(ψ, out)

Compute the partial trace of a state Ket or Bra ψ by tracing out the subsystems specified by out. Returns a density matrix. Multiple subsystems can be traced out by passing a sorted tuple of subsystem indices.

Example

julia> Φ₊ = normalize!(basis(2,0)⊗basis(2,0) + basis(2,1)⊗basis(2,1)) # Bell pair
4-d Schrodinger.Ket{SparseVector{Float64,Int64},2} with space dimensions 2⊗2:
0.71∠0°|0,0⟩ + 0.71∠0°|1,1⟩

julia> ptrace(Φ₊,1) # trace out qubit 1
2×2 Schrodinger.Density{Array{Float64,2},1} with space dimensions 2:
 0.5  0.0
 0.0  0.5

source