Operator Library

Operator Library

Schrodinger.create — Function.

create(N, dense=false)

Generate a quantum harmonic oscillator raising (creation) operator $\hat{a}^†$ in a truncated Hilbert space of size N.

Example

julia> create(4)
4×4 Schrodinger.Operator{SparseMatrixCSC{Float64,Int64},1} with space dimensions 4:
 0.0  0.0      0.0      0.0
 1.0  0.0      0.0      0.0
 0.0  1.41421  0.0      0.0
 0.0  0.0      1.73205  0.0

source

Schrodinger.destroy — Function.

destroy(N, dense=false)

Generate a quantum harmonic oscillator lowering (annihilation) operator $\hat{a}$ in a truncated Hilbert space of size N.

Example

julia> destroy(4)
4×4 Schrodinger.Operator{SparseMatrixCSC{Float64,Int64},1} with space dimensions 4:
 0.0  1.0  0.0      0.0
 0.0  0.0  1.41421  0.0
 0.0  0.0  0.0      1.73205
 0.0  0.0  0.0      0.0

source

Schrodinger.displacementop — Function.

displacementop(N, α, dense=true)

Generate a quantum harmonic oscillator displacement operator $\hat{D}(α)$ in a truncated Hilbert space of size N.

\[\hat{D}(α) = \exp\left(α\hat{a}^† - α^*\hat{a}\right)\]

Example

julia> displacementop(3,0.5im)
3×3 Schrodinger.Operator{Array{Complex{Float64},2},1} with space dimensions 3:
   0.88262+0.0im            0.0+0.439802im  -0.166001+0.0im
       0.0+0.439802im  0.647859+0.0im             0.0+0.621974im
 -0.166001+0.0im            0.0+0.621974im    0.76524+0.0im

source

Schrodinger.numberop — Function.

numberop(N, dense=false)

Generate a number operator $\hat{n}$ in a Hilbert space of size N.

Example

julia> numberop(4)
4×4 Schrodinger.Operator{SparseMatrixCSC{Float64,Int64},1} with space dimensions 4:
 0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  2.0  0.0
 0.0  0.0  0.0  3.0

source

Schrodinger.qeye — Function.

qeye(N, dims=(N,), dense=false)

Generate an identity operator for a Hilbert space of size N. It is possible to specify the subspace dimensions with the dims argument.

Example

julia> qeye(4,(2,2))
4×4 Schrodinger.Operator{SparseMatrixCSC{Float64,Int64},2} with space dimensions 2⊗2:
 1.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0

source

Schrodinger.qzero — Function.

qzero(N, dims=(N,), dense=false)

Generate a zero operator for a Hilbert space of size N. It is possible to specify the subspace dimensions with the dims argument.

Example

julia> qzero(4,(2,2))
4×4 Schrodinger.Operator{SparseMatrixCSC{Float64,Int64},2} with space dimensions 2⊗2:
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

source

Schrodinger.squeezeop — Function.

squeezeop(N, z, dense=true)

Generate a quantum harmonic oscillator squeeze operator $\hat{S}(z)$ in a truncated Hilbert space of size N.

\[\hat{S}(z) = \exp\left(\frac{1}{2}\left(z^*\hat{a}^2 - z\hat{a}^{†2}\right)\right)\]

Example

julia> squeezeop(3,0.5im)
3×3 Schrodinger.Operator{Array{Complex{Float64},2},1} with space dimensions 3:
 0.938148+0.0im       0.0+0.0im       0.0-0.346234im
      0.0+0.0im       1.0+0.0im       0.0+0.0im
      0.0-0.346234im  0.0+0.0im  0.938148+0.0im

source