Function Library

Function Library

Schrodinger.expect — Method.

expect(σ,ψ), expect(σ,ρ)

Compute the expectation value of an operator $σ$ given a state ket $|ψ⟩$ or a density matrix $ρ$. The expectation value is defined as

\[\begin{align*} E(σ,|ψ⟩) &= ⟨ψ|σ|ψ⟩, \\ E(σ,ρ) &= \textrm{tr}(σρ). \end{align*} A specialized method exists for vector of `Ket` or `Operator` inputs.\]

source

Schrodinger.fidelity — Method.

fidelity(ρ,σ), fidelity(ρ,ψ), fidelity(ψ,ϕ)

Compute the fidelity between density matrices $ρ$ and $σ$, a density matrix $ρ$ and a ket $|ψ⟩$, or two kets $|ψ⟩$ and $|ϕ⟩$. The fidelity in those three cases is defined as

\[\begin{align*} F(ρ,σ) &= \textrm{tr}\sqrt{ρ^{1/2}σρ^{1/2}}, \\ F(ρ,|ψ⟩) &= \sqrt{⟨ψ|ρ|ψ⟩}, \\ F(|ψ⟩,|ϕ⟩) &= \left|⟨ψ|ϕ⟩\right|. \end{align*}\]

See also fidelity2, which is the square of the state fidelity.

source

Schrodinger.fidelity2 — Method.

fidelity2(ρ,ψ)

Compute the Uhlmann state fidelity between density matrices $ρ$ and $σ$, a density matrix $ρ$ and a ket $|ψ⟩$, or two kets $|ψ⟩$ and $|ϕ⟩$. The Uhlmann state fidelity is simply defined as the square of the "regular" state fidelity.

source

Schrodinger.levelprobs — Method.

levelprobs(ψ), levelprobs(ψ,s)

Compute the level occupation probabilities. For a Ket, this simply corresponds to the absolute square of the amplitude of each level. For an Operator, the function returns the diagonal.

A system index, or vector of indices, can be passed as a second argument. In that case, the full system will first be partial traced to keep only the desired index. Level occupation probabilities are then calculated from the resulting reduced density matrix. If a vector of indices is passed, occupation probabilities are calculated for a fully reduced density matrix for each index.

A specialized method exists for vector of Ket or Operator inputs.

source

Schrodinger.ptrace — Method.

ptrace(ψ, out)

Compute the partial trace of a state Ket or Bra ψ by tracing out the subsystems specified by out. Returns a density matrix. Multiple subsystems can be traced out by passing a sorted tuple of subsystem indices.

Example

julia> Φ₊ = normalize!(basis(2,0)⊗basis(2,0) + basis(2,1)⊗basis(2,1)) # Bell pair
4-d Schrodinger.Ket{SparseVector{Float64,Int64},2} with space dimensions 2⊗2:
0.71∠0°|0,0⟩ + 0.71∠0°|1,1⟩

julia> ptrace(Φ₊,1) # trace out qubit 1
2×2 Schrodinger.Operator{Array{Float64,2},1} with space dimensions 2:
 0.5  0.0
 0.0  0.5

source

Schrodinger.ptrace — Method.

ptrace(ρ, out)

Compute the partial trace of a linear Operator ρ by tracing out the subsystems specified by out. Multiple subsystems can be traced out by passing a sorted tuple of subsystem indices.

Example

Φ₊ = normalize!(basis(2,0)⊗basis(2,0) + basis(2,1)⊗basis(2,1)) # Bell pair
Ψ₊ = normalize!(basis(2,0)⊗basis(2,1) + basis(2,1)⊗basis(2,0)) # Bell pair
ρ  = 0.25 * Operator(Φ₊) + 0.75 * Operator(Ψ₊) # density matrix
ptrace(ρ,2) # trace out qubit 2
# output
2×2 Schrodinger.Operator{Array{Float64,2},1} with space dimensions 2:
 0.5  0.0
 0.0  0.5

source