Function Library

Schrodinger.entanglement_fidelity — Function

entanglement_fidelity(U,V)

Compute the entanglement fidelity of an imperfect quantum operation V with respect to an ideal operation U. The entanglement fidelity is given by the following formula:

\[F_\text{e}(\mathcal{E}) = ⟨ϕ|(I⊗\mathcal{E})(ϕ)|ϕ⟩,\]

where $|ϕ⟩ = \frac{1}{\sqrt{d}}\sum_{i=1}^d|i⟩|i⟩$ is the maximally entangled state and $\mathcal{E}$ is a trace preserving quantum operation.

When trying to find the fidelity of a quantum gate $U$, implemented imperfectly as $V$, then $\mathcal{E} = U^†∘V$. In the case where $U$ and $V$ are a simple operators,

\[\begin{align*} (I⊗\mathcal{E})(ϕ) = (I⊗U^†V)|ϕ⟩⟨ϕ|(I⊗U^†V)^† \quad\text{and so} \end{align*}\]

\[\begin{align*} F_\text{e}(\mathcal{E}) &= ⟨ϕ|(I⊗U^†V)|ϕ⟩⟨ϕ|(I⊗U^†V)^†|ϕ⟩ \\ &= |⟨ϕ|(I⊗U^†V)|ϕ⟩|^2 \\ &= |⟨U,V⟩|^2/d^2 \\ \end{align*}\]

The entanglement fidelity is related to the average gate fidelity by

\[\begin{align*} \overline{F}(\mathcal{E}) = \frac{d F_\text{e}(\mathcal{E}) + 1}{d + 1}. \end{align*}\]

Schrodinger.expect — Method

expect(σ,ψ), expect(σ,ρ)

Compute the expectation value of an operator $σ$ given a state ket $|ψ⟩$ or a density matrix $ρ$. The expectation value is defined as

\[\begin{align*} E(σ,|ψ⟩) &= ⟨ψ|σ|ψ⟩, \\ E(σ,ρ) &= \textrm{tr}(σρ). \end{align*}\]

A specialized method exists for vector of Ket or Operator inputs.

Schrodinger.fidelity — Method

fidelity(ρ,σ), fidelity(ρ,ψ), fidelity(ψ,ϕ)

Compute the fidelity between density matrices $ρ$ and $σ$, a density matrix $ρ$ and a ket $|ψ⟩$, or two kets $|ψ⟩$ and $|ϕ⟩$. The fidelity in those three cases is defined as

\[\begin{align*} F(ρ,σ) &= \textrm{tr}\sqrt{ρ^{1/2}σρ^{1/2}}, \\ F(ρ,|ψ⟩) &= \sqrt{⟨ψ|ρ|ψ⟩}, \\ F(|ψ⟩,|ϕ⟩) &= \left|⟨ψ|ϕ⟩\right|. \end{align*}\]

See also fidelity2, which is the square of the state fidelity.

Schrodinger.fidelity2 — Method

fidelity2(ρ,ψ)

Compute the Uhlmann state fidelity between density matrices $ρ$ and $σ$, a density matrix $ρ$ and a ket $|ψ⟩$, or two kets $|ψ⟩$ and $|ϕ⟩$. The Uhlmann state fidelity is simply defined as the square of the "regular" state fidelity.

Schrodinger.gate_fidelity — Function

gate_fidelity(U,V)

Compute the average gate fidelity of an imperfect quantum operation V with respect to an ideal operation U.

The average gate fidelity is related to the entanglement fidelity by

\[\begin{align*} \overline{F}(\mathcal{E}) = \frac{d F_\text{e}(\mathcal{E}) + 1}{d + 1}. \end{align*}\]

Schrodinger.levelprobs — Method

levelprobs(ψ), levelprobs(ψ,s)

Compute the level occupation probabilities. For a Ket, this simply corresponds to the absolute square of the amplitude of each level. For an Operator, the function returns the diagonal.

A system index, or vector of indices, can be passed as a second argument. In that case, the full system will first be partial traced to keep only the desired index. Level occupation probabilities are then calculated from the resulting reduced density matrix. If a vector of indices is passed, occupation probabilities are calculated for a fully reduced density matrix for each index.

Specialized methods exists for vectors of Ket or Operator inputs.

Schrodinger.purity — Method

purity(ρ)

Compute the purity of a quantum state ρ, defined by

\[\begin{align*} γ = \textrm{tr}(ρ^†ρ). \end{align*}\]

This is the same as the square of the Frobenius norm of a matrix.

The purity of a quantum state is bounded below by 1/d, for a maximally mixed state, and above by 1, for a pure state. Note that the purity of a Ket is 1 by definition.

The purity is related to the linear entropy $S_L$ by $S_L = 1-γ$.

Schrodinger.ptrace — Method

ptrace(ψ, out)

Compute the partial trace of a state Ket or Bra ψ by tracing out the subsystems specified by out. Returns a density matrix. Multiple subsystems can be traced out by passing a sorted tuple of subsystem indices.

Example

julia> Φ₊ = normalize!(basis(2,0)⊗basis(2,0) + basis(2,1)⊗basis(2,1)) # Bell pair
4-d Ket{SparseVector{Float64,Int64},2} with dimensions 2⊗2
0.71∠0°|0,0⟩ + 0.71∠0°|1,1⟩

julia> ptrace(Φ₊,1) # trace out qubit 1
2×2 Operator{Array{Float64,2},1} with dimensions 2
 0.5  0.0
 0.0  0.5

Schrodinger.ptrace — Method

ptrace(ρ, out)

Compute the partial trace of a linear Operator ρ by tracing out the subsystems specified by out. Multiple subsystems can be traced out by passing a sorted tuple of subsystem indices.

Example

Φ₊ = normalize!(basis(2,0)⊗basis(2,0) + basis(2,1)⊗basis(2,1)) # Bell pair
Ψ₊ = normalize!(basis(2,0)⊗basis(2,1) + basis(2,1)⊗basis(2,0)) # Bell pair
ρ  = 0.25 * Operator(Φ₊) + 0.75 * Operator(Ψ₊) # density matrix
ptrace(ρ,2) # trace out qubit 2
# output
2×2 Operator{Array{Float64,2},1} with dimensions 2
 0.5  0.0
 0.0  0.5