Operator Library

Schrodinger.create — Method

create(N)

Generate a quantum harmonic oscillator raising (creation) operator $\hat{a}^†$ in a truncated Hilbert space of size N. Returns a sparse matrix.

Example

julia> create(4)
4×4 Operator{SparseMatrixCSC{Float64,Int64},1} with dimensions 4
 0.0  0.0      0.0      0.0
 1.0  0.0      0.0      0.0
 0.0  1.41421  0.0      0.0
 0.0  0.0      1.73205  0.0

Schrodinger.destroy — Method

destroy(N)

Generate a quantum harmonic oscillator lowering (annihilation) operator $\hat{a}$ in a truncated Hilbert space of size N. Returns a sparse matrix.

Example

julia> destroy(4)
4×4 Operator{SparseMatrixCSC{Float64,Int64},1} with dimensions 4
 0.0  1.0  0.0      0.0
 0.0  0.0  1.41421  0.0
 0.0  0.0  0.0      1.73205
 0.0  0.0  0.0      0.0

Schrodinger.displacementop — Method

displacementop(N, α)

Generate a quantum harmonic oscillator displacement operator $\hat{D}(α)$ in a truncated Hilbert space of size N. Returns a dense matrix.

\[\hat{D}(α) = \exp\left(α\hat{a}^† - α^*\hat{a}\right)\]

Example

julia> displacementop(3,0.5im)
3×3 Operator{Array{Complex{Float64},2},1} with dimensions 3
   0.88262+0.0im            0.0+0.439802im  -0.166001+0.0im
       0.0+0.439802im  0.647859+0.0im             0.0+0.621974im
 -0.166001+0.0im            0.0+0.621974im    0.76524+0.0im

Schrodinger.numberop — Method

numberop(N)

Generate a number operator $\hat{n}$ in a Hilbert space of size N. Returns a sparse matrix.

Example

julia> numberop(4)
4×4 Operator{SparseMatrixCSC{Float64,Int64},1} with dimensions 4
 0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  2.0  0.0
 0.0  0.0  0.0  3.0

Schrodinger.projectorop — Method

projectorop(N,S)

Generate a projector on the subspaces defined by an integer or a vector/range of integers S:

\[P = \sum_{i∈S} |i⟩⟨i|.\]

Example

julia> projectorop(5,[1,3])
5×5 Operator{SparseMatrixCSC{Float64,Int64},1} with dimensions 5
 0.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  0.0
 0.0  0.0  0.0  0.0  0.0

Schrodinger.qeye — Method

qeye(N, dims=(N,))

Generate an identity operator for a Hilbert space of size N. It is possible to specify the subspace dimensions with the dims argument. Returns a sparse matrix.

Example

julia> qeye(4,(2,2))
4×4 Operator{SparseMatrixCSC{Float64,Int64},2} with dimensions 2⊗2
 1.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0

Schrodinger.qzero — Method

qzero(N, dims=(N,))

Generate a zero operator for a Hilbert space of size N. It is possible to specify the subspace dimensions with the dims argument. Returns a sparse matrix.

Example

julia> qzero(4,(2,2))
4×4 Operator{SparseMatrixCSC{Float64,Int64},2} with dimensions 2⊗2
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

Schrodinger.squeezeop — Method

squeezeop(N, z)

Generate a quantum harmonic oscillator squeeze operator $\hat{S}(z)$ in a truncated Hilbert space of size N. Returns a dense matrix.

\[\hat{S}(z) = \exp\left(\frac{1}{2}\left(z^*\hat{a}^2 - z\hat{a}^{†2}\right)\right)\]

Example

julia> squeezeop(3,0.5im)
3×3 Operator{Array{Complex{Float64},2},1} with dimensions 3
 0.938148+0.0im       0.0+0.0im       0.0-0.346234im
      0.0+0.0im       1.0+0.0im       0.0+0.0im
      0.0-0.346234im  0.0+0.0im  0.938148+0.0im

Schrodinger.rand_unitary — Function

rand_unitary(N, dims=(N,))

Generate a Haar distributed random unitary operator for a Hilbert space of size N. It is possible to specify the subspace dimensions with the dims argument. Returns a dense matrix.

Example

julia> U = rand_unitary(4,(2,2));

julia> U'*U ≈ qeye(4,(2,2))
true