Quantum Object Types

Quantum Object Types

Schrodinger.Bra — Type.

Bra(x, dims=(length(x),))

Bra vector type. The dual vector to the Ket.

The Bra type has two fields, data and dims, which store the vector data and the subspace dimensions. A Bra, like a Density matrix or and Operator is parameterized by the number of subspaces it lives in. Two different kets must have the same system dimensions in order to be added together.

It is possible to normalize the bra vector after construction with the normalize! function.

source

Schrodinger.Density — Type.

Density(A, dims=(size(A,1),))

Construct a density matrix (a.k.a. density operator) from the Hermitian matrix A. An N×N matrix will by default be assumed to describe a single Hilbert space of dimension N. If the matrix represents a tensor product of Hilbert spaces, the dimensions can be defined manually by passing a tuple of subspace dimensions dims. In that case, prod(dims) must equal size(A,1). By default, the matrix is stored in sparse format.

The Density type has two fields, data and dims, which store the matrix data and the subspace dimensions. A Density matrix, like a Ket or an Operator, is parameterized by the number of subspaces it lives in. Two different density matrices must have the same system dimensions in order to be added together. A Density matrix is always Hermitian, this is enforced on construction by "hermitianizing" the given matrix.

It is possible to normalize the density matrix after construction with the normalize! function.

Example

julia> A = [1 5 2; 5 1 0; 2 0 2]
3×3 Array{Int64,2}:
 1  5  2
 5  1  0
 2  0  2
julia> σ = normalize!(Density(A))
3×3 Schrodinger.Density{Array{Float64,2},1} with space dimensions 3:
 0.25  1.25  0.5
 1.25  0.25  0.0
 0.5   0.0   0.5
julia> trace(σ)
1.0

source

Schrodinger.Ket — Type.

Ket(x, dims=(length(x),))

Construct a ket state vector from the vector x. A vector of length N will by default be assumed to be an element of a single Hilbert space of dimension N. If the vector is an element of a tensor product of Hilbert spaces, the dimensions can be defined manually by passing a tuple of subspace dimensions dims. In that case, prod(dims) must equal length(x). By default, the vector is stored in sparse format.

The Ket type has two fields, data and dims, which store the vector data and the subspace dimensions. A Ket, like a Density matrix or and Operator is parameterized by the number of subspaces it lives in. Two different kets must have the same system dimensions in order to be added together.

It is possible to normalize the ket vector after construction with the normalize! function.

Example

julia> ψ = normalize!(Ket([1,1]))
2-d Schrodinger.Ket{Array{Float64,1},1} with space dimensions 2:
0.71∠0°|0⟩ + 0.71∠0°|1⟩

source

Schrodinger.Operator — Type.

Operator(B, dims=(size(B,1),))

Construct a linear operator from the matrix B. An N×N matrix will by default be assumed to describe an operator that acts on a single Hilbert space of dimension N. If the matrix represents a linear operator on a tensor product of Hilbert spaces, the dimensions can be defined manually by passing a tuple of subspace dimensions dims. In that case, prod(dims) must equal size(B,1).

The Operator type has two fields, data and dims, which store the matrix data and the subspace dimensions. An Operator, like a Ket or a Density matrix, is parameterized by the number of subspaces it lives in. Two different density matrices must have the same system dimensions in order to be added together. An Operator may or may not be Hermitian.

Example

julia> σ = Operator([0 -im ; im 0])
2×2 Schrodinger.Operator{Array{Complex{Float64},2},1} with space dimensions 2:
 0.0+0.0im  0.0-1.0im
 0.0+1.0im  0.0+0.0im

source