State Library

State Library

Schrodinger.basis — Method.

basis(N, n)

Generate a basis state (a.k.a. Fock or number state) ket $|n⟩$, in a Hilbert space of size N. Note that the size of the Hilbert space must be at least n+1. The function fock is an alias for basis.

Returns a sparse vector.

Example

julia> ψ = basis(3,2)
3-d Schrodinger.Ket{SparseVector{Float64,Int64},1} with space dimensions 3:
1.00∠0°|2⟩

source

Schrodinger.coherent — Function.

coherent(N, α, analytic=false)

Generate a coherent state ket $|α⟩$, in a Hilbert space of size N. To create a coherent density operator, use the Operator function: Operator(coherent(N,n)).

Two methods can be used for generating a coherent state: via application of a displacment operator on a ground state (the default), or analytically, with the formula

\[|α⟩ = e^{-\frac{|α|^2}{2}} \sum_{n=0}^{N-1} \frac{α^n}{\sqrt{n!}} |n⟩.\]

While the operator method will return a normalized ket, the analytic method will not. Both methods converge as N gets larger. The analytic method is also much faster, especially for large N.

Returns a dense vector.

Example

julia> coherent(6,0.4+1im)
6-d Schrodinger.Ket{Array{Complex{Float64},1},1} with space dimensions 6:
0.60∠68°|1⟩ + 0.56∠0°|0⟩ + 0.46∠136°|2⟩ + 0.29∠-155°|3⟩ + 0.15∠-87°|4⟩

source

Schrodinger.maxentangled — Method.

maxentangled(N,n)

Generate a maximally entangled state between n N-d systems:

\[|\phi⟩=\sum_{j=0}^{N-1}\frac{1}{\sqrt{N}}|j⟩^{⊗n}.\]

source

Schrodinger.maxmixed — Method.

maxmixed(N)

Generate a maximally mixed density matrix. The maximally mixed state is a mixture of basis states with uniform probability.

Returns a sparse matrix.

Example

julia> maxmixed(4)
4×4 Schrodinger.Operator{SparseMatrixCSC{Float64,Int64},1} with space dimensions 4:
 0.25  0.0   0.0   0.0
 0.0   0.25  0.0   0.0
 0.0   0.0   0.25  0.0
 0.0   0.0   0.0   0.25

source

Schrodinger.thermal — Method.

thermal(N, n)

Generate a thermal state density matrix $ρ_n$ with particle number n, in a Hilbert space of size N. A thermal state $ρ_n$ is a probabilistic mixture of basis states such that the expectation value of the number operator $\hat{n}$ is n. Note that this is true only if $N≫n$. The returned density matrix is always normalized.

Returns a sparse matrix.

Example

julia> N=5; n=0.2;

julia> ρ = thermal(N,n)
5×5 Schrodinger.Operator{SparseMatrixCSC{Float64,Int64},1} with space dimensions 5:
 0.833441  0.0       0.0        0.0         0.0
 0.0       0.138907  0.0        0.0         0.0
 0.0       0.0       0.0231511  0.0         0.0
 0.0       0.0       0.0        0.00385852  0.0
 0.0       0.0       0.0        0.0         0.000643087

julia> expect(numberop(N),ρ)
0.19935691318327978

source